Rename yacc.rs to yacc_parser.rs.
This is, hopefully, a less confusing name.
